Parse the target path given to cd to prevent recursive loop.
If cd was given {release_path} as target path, it would attempt to run a command to figure out what the value of {release_path} is, which, in the process, would also want to know the value of {release_path} and so on.

The fix simply parses the value as it is given, thus making the result static instead of a closure.