Fix bug #555 and clean output
The bug #555 was caused by using a wrong escape-sequence for the
terminal.  Instead of using \e[K it has used \e[k which actually has no
meaning and only bloats some noise on the terminal (except on OSX where
the noise is catched and hidden from the user). To see this behaviour
you need a real linux client for example a VM and you have to use it's
own terminal – not the OSX-terminal via ssh or something like that. Than
type the following example to see the behavior:

    echo -e "aaaaaaaaa\r\033[k"
    echo -e "bbbbbbbbb\r\033[K"

Furthermore the cursor is now set to the beginning of the line first to
really clean the line and not just something starting from a random
position.