Run deploy:writable again after cache warmup
As described in ticket #575 the cache is not writable after the
deployment of symfony if the deployment-user is not the httpd-user.
Therefore we would make the cache-directory writable again after the
cache-warmup.  This is necessary because symfony may not do that and the
deploy user should know who is able to write into the cache.  You can
finally define this by using `set('writable_dirs', […])` and
`set('http_user', '…')`.