Let rsync honor 'become' option for host
If a host has a 'become' option set, it seems sensible that rsync uploads
should also be performed as that same user, which is currently not the case.
This change makes rsync run as the 'become' user by running sudo on the
remote side, just like how 'become' is implemented for normal SSH commands.