Always update shared folder from git
Right now `deployer shared` would install the shared directory
once; when the first deployment of e.g. a branch is running.

Adding items to a shared resource in git will not update the
deployments shared folder. This is often confusing for our devs
who are required to add some assets sometimes.

I'd suggest this simple "fix" to always update the shared folder
from the git repository.

Signed-off-by: Mark Kusch <mark.kusch@transfermarkt.de>