Add tasks to cache a manifest of event listeners
Laravel supports automatic discovery of events and listeners since
version 5.8.9. (https://laravel.com/docs/5.8/events#event-discovery)
As the documentation states, you want to cache the manifest of event
listeners in a production environment to prevent scanning the
directories on every request.

The new tasks are not added to the deploy task because automatic event
discovery is opt-in functionality in Laravel.