[issue-2176] Added `escape`/`unescape` functions
This commit adds two new functions:
* `escape(string): string`
* `unescape(string): string`

Once the value has been processed by the `escape` function, then
the values wrapped in `{{}}` like `{{foo}}` will not be parsed by
the `parse` function.

The `escape` function works by replacing the `{{` with `\\{\\}` and
`}}` with `\\}\\}`. The `unescape` function reverses the process.

The `parse` function has been updated to run the `unescape` after
parsing has been completed, to ensure that the commands are being
executed normally.

Fixes #2176