[issue-2187] Type declarations for src/Configuration
Refactored the code in `src/Configuration` to include type declarations
for method arguments and returned values.

To make it easier to track what's left to do and to have a tool that
prevents "regressions" in the future, PHP_CodeSniffer and Slevomat
Coding Standard have been added.

Next, a `phpcs.xml.dist` file has been created and configured to
enforce type hints for methods. For the time being, only files in
the src/Configuration directory are now being watched.

To make it simpler to execute the checks, `lint` script has been
added to `composer.json` to allow running checks with `composer lint`.