DRAFT: First activate plugins/themes THEN build them (#2607)
* First activate plugins/themes THEN build them

Currently the first step is to run bin/build.sh which builds all *ACTIVATED* themes. After this step we activate all composer-registered plugins (which in my case are all the themes) and THEN compile them. For new themes that means, that they are *NOT* build because they are activated AFTER building.

This PR fixes that and activated first and builds second.

* update changelog