feat: add support for hidden keys in yaml files
During YAML loading the aliases are already resolved so there is no point in validating them or trying to execute them during import. That's why I've chosen to remove them completely after parsing.

I had to add the "# test.yaml" comment at the end because the importer checks for this, assuming a file will be passed to it. However, I wasn't sure where to put the file and keeping it isolated to the test seemed fine as well.