test: restore the Deployer input/output after test
This is necessary because the static instance gets set somewhere before TaskTest.php is executed, but gets overridden for the purposes of this test. Simply unsetting the deployer causes the instance to get garbage collected, which causes TaskTest.php to fail because it no longer exists.

I didn't really want to touch other code since it'd become pretty massive in scope, I think. As such, this was the quickest way to work around the current design of the code and how it interacts with tests.

Also, this may cause issues in other tests because the hosts I'm importing in this test will remain in the container, but I couldn't figure out how to remove hosts again after they've already been set. I tried extending the container with a new "hosts" value but that gave me a FrozenServiceException, as did directly overriding it, so I'm at a bit of a loss here. :)